requirements.txt
openai-whisper
sounddevice
numpy
groq
elevenlabs
torch